+++ cmd/lambda/update_redbox_principal_policy/main.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/Optum/Redbox/pkg/api/response"
+	"github.com/Optum/Redbox/pkg/awsiface"
 	"github.com/Optum/Redbox/pkg/common"
 	"github.com/Optum/Redbox/pkg/db"
 	"github.com/Optum/Redbox/pkg/rolemanager"
@@ -27,12 +28,13 @@ func main() {
 
 func handler(ctx context.Context, snsEvent events.SNSEvent) error {
 	dbSvc, err := db.NewFromEnv()
-	awsSession := newAWSSession()
+	awsSession := session.Must(session.NewSession())
 	tokenSvc := common.STS{Client: sts.New(awsSession)}
 	s3Svc := common.S3{
 		Client:  s3.New(awsSession),
 		Manager: s3manager.NewDownloader(awsSession),
 	}
+	roleManagerSvc := &rolemanager.IAMPolicyManager{}
 
 	if err != nil {
 		log.Printf("Unable to setup DB Service: %s", err.Error())
@@ -48,50 +50,77 @@ func handler(ctx context.Context, snsEvent events.SNSEvent) error {
 			log.Printf("Failed to read SNS message %s: %s", snsRecord.Message, err.Error())
 			return err
 		}
-		account, err := dbSvc.GetAccount(lease.AccountID)
-		if err != nil {
-			log.Printf("Failed to get the redbox account %s: %s",
-				lease.AccountID, err.Error())
-			return err
-		}
 
-		accountRes := response.AccountResponse(*account)
-		fmt.Printf("AccountID = %s \nRoleToAssume = %s\n", lease.AccountID, accountRes.AdminRoleArn)
-		principalIAMDenyTags := strings.Split(common.RequireEnv("PRINCIPAL_IAM_DENY_TAGS"), ",")
-		principalRoleName := common.RequireEnv("PRINCIPAL_ROLE_NAME")
-		policyName := common.RequireEnv("PRINCIPAL_POLICY_NAME")
-		principalPolicyArn, err := arn.Parse(fmt.Sprintf("arn:aws:iam::%s:policy/%s", lease.AccountID, policyName))
-		if err != nil {
-			log.Printf("Failed to parse ARN 'arn:aws:iam::%s:policy/%s': %s", lease.AccountID, policyName, err.Error())
-			return err
-		}
-
-		policy, err := getPolicy(s3Svc, getPolicyInput{
-			PrincipalPolicyArn:   principalPolicyArn.String(),
-			PrincipalRoleArn:     fmt.Sprintf("arn:aws:iam::%s:role/%s", lease.AccountID, principalRoleName),
-			PrincipalIAMDenyTags: principalIAMDenyTags,
-			AdminRoleArn:         accountRes.AdminRoleArn,
-		})
-
-		// Assume role into the new Redbox account
-		accountSession, err := tokenSvc.NewSession(awsSession, accountRes.AdminRoleArn)
-		if err != nil {
-			log.Printf("Failed to assume role '%s': %s", accountRes.AdminRoleArn, err.Error())
-			return err
-		}
-		iamSvc := iam.New(accountSession)
-
-		// Create the Role + Policy
-		roleManager := &rolemanager.IAMPolicyManager{}
-		roleManager.SetIAMClient(iamSvc)
-		return roleManager.MergePolicy(&rolemanager.MergePolicyInput{
-			PolicyName:        policyName,
-			PolicyArn:         principalPolicyArn,
-			PolicyDocument:    policy,
-			PolicyDescription: "", // Policy should already exist so this will be ignored
+		err = processRecord(processRecordInput{
+			AccountID:            lease.AccountID,
+			DbSvc:                dbSvc,
+			StoragerSvc:          s3Svc,
+			TokenSvc:             tokenSvc,
+			AwsSession:           awsSession,
+			RoleManager:          roleManagerSvc,
+			PrincipalRoleName:    common.RequireEnv("PRINCIPAL_ROLE_NAME"),
+			PrincipalPolicyName:  common.RequireEnv("PRINCIPAL_POLICY_NAME"),
+			PrincipalIAMDenyTags: strings.Split(common.RequireEnv("PRINCIPAL_IAM_DENY_TAGS"), ","),
+			PolicyBucket:         common.RequireEnv("ARTIFACTS_BUCKET"),
+			PolicyBucketKey:      common.RequireEnv("PRINCIPAL_POLICY_S3_KEY"),
 		})
 	}
 	return nil
+}
+
+type processRecordInput struct {
+	AccountID            string
+	DbSvc                db.DBer
+	StoragerSvc          common.Storager
+	TokenSvc             common.TokenService
+	AwsSession           awsiface.AwsSession
+	RoleManager          rolemanager.PolicyManager
+	PrincipalRoleName    string
+	PrincipalPolicyName  string
+	PrincipalIAMDenyTags []string
+	PolicyBucket         string
+	PolicyBucketKey      string
+}
+
+func processRecord(input processRecordInput) error {
+
+	account, err := input.DbSvc.GetAccount(input.AccountID)
+	if err != nil {
+		log.Printf("Failed to get the redbox account %s: %s",
+			input.AccountID, err.Error())
+		return err
+	}
+
+	accountRes := response.AccountResponse(*account)
+
+	principalPolicyArn, err := arn.Parse(fmt.Sprintf("arn:aws:iam::%s:policy/%s", input.AccountID, input.PrincipalPolicyName))
+	if err != nil {
+		log.Printf("Failed to parse ARN 'arn:aws:iam::%s:policy/%s': %s", input.AccountID, input.PrincipalPolicyName, err.Error())
+		return err
+	}
+
+	policy, err := input.StoragerSvc.GetTemplateObject(input.PolicyBucket, input.PolicyBucketKey, getPolicyInput{
+		PrincipalPolicyArn:   principalPolicyArn.String(),
+		PrincipalRoleArn:     fmt.Sprintf("arn:aws:iam::%s:role/%s", input.AccountID, input.PrincipalRoleName),
+		PrincipalIAMDenyTags: input.PrincipalIAMDenyTags,
+		AdminRoleArn:         accountRes.AdminRoleArn,
+	})
+
+	// Assume role into the new Redbox account
+	accountSession, err := input.TokenSvc.NewSession(input.AwsSession, accountRes.AdminRoleArn)
+	if err != nil {
+		log.Printf("Failed to assume role '%s': %s", accountRes.AdminRoleArn, err.Error())
+		return err
+	}
+	iamSvc := iam.New(accountSession)
+
+	// Create the Role + Policy
+	input.RoleManager.SetIAMClient(iamSvc)
+	return input.RoleManager.MergePolicy(&rolemanager.MergePolicyInput{
+		PolicyName:     input.PrincipalPolicyName,
+		PolicyArn:      principalPolicyArn,
+		PolicyDocument: policy,
+	})
 }
 
 type getPolicyInput struct {
@@ -99,20 +128,4 @@ type getPolicyInput struct {
 	PrincipalRoleArn     string
 	PrincipalIAMDenyTags []string
 	AdminRoleArn         string
-}
-
-func getPolicy(storage common.S3, input getPolicyInput) (string, error) {
-	bucket := common.RequireEnv("ARTIFACTS_BUCKET")
-	key := common.RequireEnv("PRINCIPAL_POLICY_S3_KEY")
-	policy, err := storage.GetTemplateObject(bucket, key, input)
-	return policy, err
-}
-
-func newAWSSession() *session.Session {
-	awsSession, err := session.NewSession()
-	if err != nil {
-		errorMessage := fmt.Sprintf("Failed to create AWS session: %s", err)
-		log.Fatal(errorMessage)
-	}
-	return awsSession
 }